Handle address overflow in findMatchingRegion()
Previously attempting a read/write from 0xFFFFFFFC would match any
region since (address + size) == 0.  I added a few unit tests to
MemorySimTest.cpp to expose this and related overflow issues.

+++ libmemsim/src/MemorySim.c
@@ -183,7 +183,7 @@ static MemoryRegion* findMatchingRegion(MemorySim* pThis, uint32_t address, uint
     while (pCurr)
     {
         MemoryRegion* pNext = pCurr->pNext;
-        if (address >= pCurr->baseAddress && address + size <= pCurr->baseAddress + pCurr->size)
+        if (address >= pCurr->baseAddress && (uint64_t)address + size <= (uint64_t)pCurr->baseAddress + pCurr->size)
             return pCurr;
         pCurr = pNext;
     }
